+++ packages/core/src/components/AlertBanner/__stories__/AlertBanner.stories.tsx
@@ -4,7 +4,7 @@ import AlertBannerText from "../AlertBannerText/AlertBannerText";
 import AlertBannerLink from "../AlertBannerLink/AlertBannerLink";
 import AlertBannerButton from "../AlertBannerButton/AlertBannerButton";
 import { createStoryMetaSettingsDecorator } from "../../../storybook/functions/createStoryMetaSettingsDecorator";
-import "./AlertBanner.stories.scss";
+import Flex from "../../Flex/Flex";
 import { Meta, StoryObj } from "@storybook/react";
 
 const metaSettings = createStoryMetaSettingsDecorator({
@@ -30,23 +30,26 @@ export default {
     AlertBannerButton
   },
   argTypes: metaSettings.argTypes,
-  decorators: metaSettings.decorators
-} satisfies Meta<typeof AlertBanner>;
+  decorators: [
+    ...metaSettings.decorators,
+    (Story: React.ComponentType) => (
+      <div style={{ width: "610px" }}>
+        <Story />
+      </div>
+    )
+  ]
+} as Meta<typeof AlertBanner>;
 
 export const Overview = {
   render: alertBannerTemplate.bind({}),
-  name: "Overview",
-
-  args: {
-    className: "monday-storybook-alert-banner_big-container"
-  }
+  name: "Overview"
 };
 
 type Story = StoryObj<typeof AlertBanner>;
 
 export const Types: Story = {
   render: () => (
-    <div className="monday-storybook-alert-banner_column-wrapper monday-storybook-alert-banner_big-container">
+    <Flex direction="column" gap={16}>
       <AlertBanner>
         <AlertBannerText text="Alert banner message" />
         <AlertBannerLink text="this is a CTA" href="https://monday.com" />
@@ -67,63 +70,64 @@ export const Types: Story = {
         <AlertBannerText text="Alert banner message" />
         <AlertBannerLink text="this is a CTA" href="https://monday.com" />
       </AlertBanner>
-    </div>
+    </Flex>
   ),
-
   name: "Types"
 };
 
 export const AlertBannerWithButton: Story = {
   render: () => (
-    <AlertBanner className="monday-storybook-alert-banner_big-container">
+    <AlertBanner>
       <AlertBannerText text="Lorem ipsum dolor sit amet" />
       <AlertBannerButton>Lorem Ipsum</AlertBannerButton>
     </AlertBanner>
   ),
-
   name: "Alert Banner with button"
 };
 
 export const AlertBannerWithLink: Story = {
   render: () => (
-    <AlertBanner className="monday-storybook-alert-banner_big-container">
+    <AlertBanner>
       <AlertBannerText text="Alert banner message" />
       <AlertBannerLink text="this is a CTA" href="https://monday.com" />
     </AlertBanner>
   ),
-
   name: "Alert Banner with link"
 };
 
 export const AlertBannerAsAnAnnouncement: Story = {
   render: () => (
-    <AlertBanner backgroundColor="dark" className="monday-storybook-alert-banner_big-container">
+    <AlertBanner backgroundColor="dark">
       <AlertBannerText text="Join us at Elevate 2022" />
       <AlertBannerLink text="RSVP now" href="https://monday.com" />
     </AlertBanner>
   ),
-
   name: "Alert banner as an announcement"
 };
 
 export const AlertBannerAsAnOpportunityToUpgrade: Story = {
   render: () => (
-    <AlertBanner className="monday-storybook-alert-banner_big-container">
+    <AlertBanner>
       <AlertBannerText text="7 days left on your monday CRM trial" />
       <AlertBannerLink text="Upgrade now" href="https://monday.com" />
     </AlertBanner>
   ),
-
   name: "Alert banner as an opportunity to upgrade"
 };
 
 export const OverflowText: Story = {
   render: () => (
-    <AlertBanner className="monday-storybook-alert-banner_small-container">
+    <AlertBanner>
       <AlertBannerText text="This is a really long alert..." />
       <AlertBannerLink text="Call to action" href="https://monday.com" />
     </AlertBanner>
   ),
-
-  name: "Overflow text"
+  name: "Overflow text",
+  decorators: [
+    (Story: React.ComponentType) => (
+      <div style={{ width: "320px" }}>
+        <Story />
+      </div>
+    )
+  ]
 };